pyproject.toml
[tool.poetry]
name = "osint_mcp_experiment"
version = "0.1.0"
description = "OSINT, LLM, NLP, Vector Search, Geo and Dash Frontend"
authors = ["Sven Steinbauer"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

# --- LLM APIs ---
openai = "^1.35.0"
anthropic = "^0.30.0"

# --- MCP ---
mcp = "1.25.0"

# --- Google APIs ---
google-api-python-client = "^2.137.0"

# --- Vector Store ---
faiss-cpu = "^1.8.0"

# --- NLP ---
spacy = "^3.7.4"
nltk = "^3.9.1"
sentence-transformers = "^2.7.0"

# --- ML / Data ---
numpy = "^1.26.4"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"

# --- Geospatial ---
shapely = "^2.0.4"
h3 = "^4.1.0"
geopandas = "^0.14.4"

# --- Knowledge Graphs ---
networkx = "^3.3"
pyvis = "^0.3.2"

# --- Dash Frontend ---
dash = "^2.17.1"
dash-bootstrap-components = "^1.6.0"
dash-ag-grid = "^31.2.0"
dash-extensions = "^1.0.14"
dash-leaflet = "^1.0.15"
dash-trich-components = "^1.0.0"

# --- Backend API ---
fastapi = "^0.111.0"
uvicorn = ">=0.31.1"

# --- Utilities ---
python-dotenv = "^1.0.1"
rank-bm25 = "^0.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
